tidy up of application module, rename files for consistency
