+++ src/core/validation/constraints/index.ts
@@ -1,2 +1,1 @@
 export * from './unique.text.id';
-export * from './user.exists.constraint';
