+++ src/domain/community/application/index.ts
@@ -1,3 +1,3 @@
 export * from './application.entity';
-export * from './application.factory';
+export * from './application.factory.service';
 export * from './application.dto';
