+++ src/domain/community/community/community.module.ts
@@ -3,6 +3,7 @@ import { UserGroupModule } from '@domain/community/user-group/user-group.module'
 import { UserModule } from '@domain/community/user/user.module';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { ApplicationModule } from '../application/application.module';
 import { Community } from './community.entity';
 import { CommunityResolverFields } from './community.resolver.fields';
 import { CommunityResolverMutations } from './community.resolver.mutations';
@@ -13,6 +14,7 @@ import { CommunityService } from './community.service';
     UserGroupModule,
     UserModule,
     ApplicationFactoryModule,
+    ApplicationModule,
     TypeOrmModule.forFeature([Community]),
   ],
   providers: [
